backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
anthropic>=0.28.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=23.2.1
boto3>=1.34.0
google-genai>=0.7.0
python-multipart>=0.0.9
